+++ .eslintrc
@@ -6,3 +6,6 @@ env:
 
 extends:
   "htmlacademy/es5.5"
+
+parserOptions:
+  sourceType: module
