+++ index.html
@@ -254,7 +254,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="./js/main.js"></script>
 
 </body>
