+++ package.json
@@ -11,13 +11,15 @@
     "url": "https://github.com/htmlacademy-javascript/1394821-keksobooking-21/issues"
   },
   "devDependencies": {
+    "browser-sync": "^2.26.13",
     "eslint": "7.1.0",
     "eslint-config-htmlacademy": "0.6.0"
   },
   "license": "MIT",
   "dependencies": {},
   "scripts": {
-    "test": "eslint js/"
+    "test": "eslint js/",
+    "start": "browser-sync --watch"
   },
   "engines": {
     "node": "8.11"
